+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Cleaned up timeline element loading/dropping to prevent some scroll-jump edge cases
+
 ### Security
 
 ## 3.0.1